Move rate_supervision and rate_mental_health into new module. Move tests into new file. Update lime.__init__ to include imports from new Transformations module.
